# swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "MentorNepal API",
    "version": "1.0.0",
    "description": "API docs for MentorNepal backend"
  },
  "servers": [
    {
      "url": "http://localhost:3001"
    }
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
    "schemas": {
...
      "AvailabilitySlot": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "mentorId": {
            "type": "string"
          },
          "dayOfWeek": {
            "type": "integer",
            "minimum": 0,
            "maximum": 6
          },
          "startTime": {
            "type": "string",
            "pattern": "^([0-1]?[0-9]|2[0-3]):[0-5][0-9]$"
          },
          "endTime": {
            "type": "string",
            "pattern": "^([0-1]?[0-9]|2[0-3]):[0-5][0-9]$"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
      "TimeSlot": {
        "type": "object",
        "properties": {
          "startTime": {
            "type": "string",
            "format": "date-time"
          },
          "endTime": {
            "type": "string",
            "format": "date-time"
          },
          "available": {
            "type": "boolean"
          }
        }
      },
...
    }
  },
  "paths": {
...
    "/api/availability/{mentorId}": {
      "get": {
        "summary": "Get mentor's weekly availability",
        "tags": [
          "Availability"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "mentorId",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "List of availability slots",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/AvailabilitySlot"
                  }
                }
              }
            }
          },
          "404": {
            "description": "Mentor not found"
          }
        }
      }
    },
    "/api/availability": {
      "put": {
        "summary": "Update current mentor's availability",
        "tags": [
          "Availability"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "slots": {
                    "type": "array",
                    "items": {
                      "type": "object",
                      "required": [
                        "dayOfWeek",
                        "startTime",
                        "endTime"
                      ],
                      "properties": {
                        "dayOfWeek": {
                          "type": "integer",
                          "minimum": 0,
                          "maximum": 6
                        },
                        "startTime": {
                          "type": "string",
                          "pattern": "^([0-1]?[0-9]|2[0-3]):[0-5][0-9]$"
                        },
                        "endTime": {
                          "type": "string",
                          "pattern": "^([0-1]?[0-9]|2[0-3]):[0-5][0-9]$"
                        }
                      }
                    }
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Availability updated"
          },
          "400": {
            "description": "Invalid slots"
          },
          "401": {
            "description": "Not authenticated"
          },
          "403": {
            "description": "User is not a mentor"
          }
        }
      }
    },
    "/api/availability/{mentorId}/open": {
      "get": {
        "summary": "Get open time slots for a specific week (cached)",
        "tags": [
          "Availability"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "mentorId",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "weekStart",
            "required": true,
            "schema": {
              "type": "string",
              "format": "date"
            },
            "description": "Week start date (ISO 8601)"
          }
        ],
        "responses": {
          "200": {
            "description": "List of open time slots",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/TimeSlot"
                  }
                }
              }
            }
          },
          "404": {
            "description": "Mentor not found"
          }
        }
      }
    },
...
  },
...
}